backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Validation & settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Misc
python-multipart==0.0.20
redis==5.2.1
slowapi==0.1.9
aiofiles==24.1.0

# AI / RAG
openai>=1.50.0
faiss-cpu>=1.8.0
numpy>=1.26.4
python-docx>=1.1.2
pdfplumber>=0.10.0
openpyxl>=3.1.0

# Google Drive
google-api-python-client>=2.100.0
google-auth>=2.23.0
